Added Data folder; dataGarbageRemover.py desc
Added Data repository to store data for dataGarbageRemover in same format as it would be used on disk.
I accidentally committed dataGarbageRemover.py without a description — this python file looks at the csv and json files in the neighbouring data folder, goes through every relevant column for each, and removes any text in any cell in that column which matches any regex in the regex.txt file. It then saves the new dataframe as a json file in the data folder.