Refactored, fixed for equal non-English words
When using this for the word similarity function, two identical texts could have a non-1 similarity score, because they contained the empty string or nonsense words like "brecd". These words were not in the wordnet corpus, so False was returned. The function now checks to see if the two words are identical before looking at the corpus.
The code has also been shortened and the main removed.

Type hints were added to monge elkan. I made sure that it would still work with the changes to the synonym checker.

Removed getTopSImilarSentence, did not fulfill requirements.